// src/app/model/reservaEspecial.ts
export interface ReservaEspecial {
    id:                    number;
    telefono:              string;
    direccion:             string;
    email:                 string;
    tipo_identificacion:   string;
    tipo_persona:          number;
    numero_identificacion: number;
    digito_verificacion:   number;
    razon_social:          string;
    primer_apellido:       string;
    segundo_apellido:      string;
    primer_nombre:         string;
    segundo_nombre:        string;
    tarjeta_profecional:   string;
    codigo_municipio:      string;
}